requirements.txt
PySimpleGUI == 4.60.4
Pillow == 9.2.0